Moves CSS write to module
