+++ .eleventy.js
@@ -1,9 +1,6 @@
-const path = require('path');
-const fs = require('fs');
 const syntaxHighlight = require('@11ty/eleventy-plugin-syntaxhighlight');
 const svgContents = require('eleventy-plugin-svg-contents');
-const paths = require('./_client/config/paths');
-const doStyles = require('./_client/config/doStyles');
+const writeCSS = require('./_client/config/writeCSS');
 const inlineContents = require('./_client/config/filter-inlineContents');
 
 module.exports = function(eleventyConfig) {
@@ -17,42 +14,9 @@ module.exports = function(eleventyConfig) {
   eleventyConfig.addWatchTarget('./_client/src/scss/');
 
   // Process files before building.
-  eleventyConfig.on('beforeBuild', () => {
-    fs.readdirSync(paths.scss)
-      // Filter out directories.
-      .filter((file) => {
-        return fs.statSync(path.resolve(paths.scss, file)).isFile();
-      })
-      .forEach( async (file) => {
-        const entry = path.resolve(paths.scss, file);
-        const basename = path.basename(entry, '.scss');
-        const outputFilname = `css/${basename}.css`;
-        const cssPath = path.resolve(paths.src, outputFilname);
+  eleventyConfig.on('beforeBuild', writeCSS);
 
-        // Create css path if it doesn't exist.
-        if (! fs.existsSync(path.dirname(cssPath))) {
-          try {
-            fs.mkdirSync(path.dirname(cssPath), { recursive: true });
-          } catch (error) {
-            console.error(`Error making directory for CSS output: ${error}`);
-          }
-        }
-
-        // Get the processed CSS. This works without resolving the path, but we'll do it anyway.
-        const processedCss = await doStyles(path.resolve(paths.projectRoot, entry));
-
-        // Write the output to disk.
-        fs.writeFileSync(
-          cssPath,
-          processedCss,
-          (error) => console.error(`Error writing generated CSS: ${error}`)
-        );
-
-        console.log('Writing', outputFilname, 'from', file);
-      });
-  });
-
-  // Add syntax highlighting.
+  // Plugins.
   eleventyConfig.addPlugin(syntaxHighlight);
   eleventyConfig.addPlugin(svgContents);
